feat: incoming visa tc10 terminado

+++ sicf-consultas-generales/src/main/java/ob/unibanca/sicf/consultasgenerales/model/incomingvisa/IncomingVisaTC10TCR0.java
@@ -29,10 +29,10 @@ public class IncomingVisaTC10TCR0 {
 	String accountNumber;
 	String accountNumberExt;
 	double destinationAmount;
-	double destCurrencyCode;
+	Integer destCurrencyCode;
 	String descDestCurrencyCode;
 	double sourceAmount;
-	double sourceCurrencyCode;
+	Integer sourceCurrencyCode;
 	String descSourceCurrencyCode;
 	
 }
