+++ sicf-consultas-generales/src/main/java/ob/unibanca/sicf/consultasgenerales/model/incomingvisa/IncomingVisaTC10TCR0Det.java
@@ -31,10 +31,10 @@ public class IncomingVisaTC10TCR0Det {
 	String accountNumber;
 	String accountNumberExt;
 	double destinationAmount;
-	double destCurrencyCode;
+	Integer destCurrencyCode;
 	String descDestCurrencyCode;
 	double sourceAmount;
-	double sourceCurrencyCode;
+	Integer sourceCurrencyCode;
 	String descSourceCurrencyCode;
 	String messageText;
 	String settlementFlag;
